+++ orienteer-users/src/main/java/org/orienteer/hook/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Hooks package for 'Orienteer Users' module
+ */
+package org.orienteer.hook;
