+++ orienteer-users/src/main/java/org/orienteer/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Models package for 'Orienteer Users' module
+ */
+package org.orienteer.model;
