+++ orienteer-users/src/main/java/org/orienteer/module/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Orienteers modules package for 'Orienteer Users' module
+ */
+package org.orienteer.module;
