+++ orienteer-users/src/main/java/org/orienteer/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Main package for 'Orienteer Users' module
+ */
+package org.orienteer;
