+++ orienteer-users/src/main/java/org/orienteer/resource/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Resources package for 'orienteer-users' module
+ */
+package org.orienteer.resource;
