+++ orienteer-users/src/main/java/org/orienteer/service/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Services package for 'orienteer-users' module
+ */
+package org.orienteer.service;
