--- orienteer-users/src/main/java/org/orienteer/users/hook/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Hooks package for 'Orienteer Users' module
- */
-package org.orienteer.users.hook;
