--- orienteer-users/src/main/java/org/orienteer/users/model/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Models package for 'Orienteer Users' module
- */
-package org.orienteer.users.model;
