--- orienteer-users/src/main/java/org/orienteer/users/module/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Orienteers modules package for 'Orienteer Users' module
- */
-package org.orienteer.users.module;
