--- orienteer-users/src/main/java/org/orienteer/users/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Main package for 'Orienteer Users' module
- */
-package org.orienteer.users;
