--- orienteer-users/src/main/java/org/orienteer/users/resource/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Resources package for 'orienteer-users' module
- */
-package org.orienteer.users.resource;
