--- orienteer-users/src/main/java/org/orienteer/users/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Services package for 'orienteer-users' module
- */
-package org.orienteer.users.service;
