--- orienteer-users/src/main/java/org/orienteer/users/util/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Utils package for 'Orienteer Users' module
- */
-package org.orienteer.users.util;
