--- orienteer-users/src/main/java/org/orienteer/users/web/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package which contains web pages for 'orienteer-users' module
- */
-package org.orienteer.users.web;
