+++ orienteer-users/src/main/java/org/orienteer/util/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Utils package for 'Orienteer Users' module
+ */
+package org.orienteer.util;
