+++ orienteer-users/src/main/java/org/orienteer/web/DefaultRegistrationPage.html
@@ -1,0 +1,42 @@
+<wicket:extend>
+    <div class="container">
+        <h3 class="text-center">
+            <strong>Registration</strong>
+        </h3>
+        <form>
+            <div class="form-group">
+                <label>
+                    <strong>First name:</strong>
+                </label>
+                <input type="text" class="form-control"/>
+            </div>
+            <div class="form-group">
+                <label>
+                    <strong>Last name:</strong>
+                </label>
+                <input type="text" class="form-control"/>
+            </div>
+            <div class="form-group">
+                <label>
+                    <strong>E-mail:</strong>
+                </label>
+                <input type="text" class="form-control"/>
+            </div>
+            <div class="form-group">
+                <label>
+                    <strong>Password:</strong>
+                </label>
+                <input type="password" class="form-control"/>
+            </div>
+            <div class="form-group">
+                <label>
+                    <strong>Confirm password:</strong>
+                </label>
+                <input type="password" class="form-control"/>
+            </div>
+            <div class="form-group mx-auto">
+                <a class="btn btn-primary" href="#">Register</a>
+            </div>
+        </form>
+    </div>
+</wicket:extend>
