+++ orienteer-users/src/main/java/org/orienteer/web/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package which contains web pages for 'orienteer-users' module
+ */
+package org.orienteer.web;
