+++ orienteer-users/src/main/resources/META-INF/services/org.apache.wicket.IInitializer
@@ -1,1 +1,1 @@
-org.orienteer.users.Initializer
+org.orienteer.Initializer
