+++ orienteer-users/src/test/java/org/orienteer/service/TestInitModule.java
@@ -1,0 +1,10 @@
+package org.orienteer.service;
+
+import com.google.inject.AbstractModule;
+
+public class TestInitModule extends AbstractModule {
+    @Override
+    protected void configure() {
+        bind(IOrienteerUsersService.class).to(TestOrienteerUserService.class).asEagerSingleton();
+    }
+}
