+++ orienteer-users/src/test/java/org/orienteer/service/TestOrienteerUserService.java
@@ -1,0 +1,11 @@
+package org.orienteer.service;
+
+import org.orienteer.model.OrienteerUser;
+
+public class TestOrienteerUserService extends OrienteerUsersService {
+
+    @Override
+    protected void notifyUserAboutRestorePassword(OrienteerUser user) {
+        // do nothing
+    }
+}
