--- orienteer-users/src/test/java/org/orienteer/users/service/TestInitModule.java
@@ -1,10 +1,0 @@
-package org.orienteer.users.service;
-
-import com.google.inject.AbstractModule;
-
-public class TestInitModule extends AbstractModule {
-    @Override
-    protected void configure() {
-        bind(IOrienteerUsersService.class).to(TestOrienteerUserService.class).asEagerSingleton();
-    }
-}
