--- orienteer-users/src/test/java/org/orienteer/users/service/TestOrienteerUserService.java
@@ -1,11 +1,0 @@
-package org.orienteer.users.service;
-
-import org.orienteer.users.model.OrienteerUser;
-
-public class TestOrienteerUserService extends OrienteerUsersService {
-
-    @Override
-    protected void notifyUserAboutRestorePassword(OrienteerUser user) {
-        // do nothing
-    }
-}
