+++ orienteer-users/src/test/resources/META-INF/services/com.google.inject.Module
@@ -1,1 +1,1 @@
-org.orienteer.users.service.TestInitModule
+org.orienteer.service.TestInitModule
